# perf(lttng): improve memory consumption