--- a/src/test/infra/lttng/test_architecture_reader_lttng.py
+++ b/src/test/infra/lttng/test_architecture_reader_lttng.py
@@ -1,4 +1,4 @@
-# Copyright 2021 esearch Institute of Systems Planning, Inc.
+# Copyright 2021 Research Institute of Systems Planning, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.