… src/test/infra/lttng/test_lttng.py