Signed-off-by: hsgwa <hasegawa.isp@gmail.com>

This PR improves processing time and memory usage for Lttng class.

For now, I removed ros2_tracing classes.
memory usage was improved, but processing time was worse than with cache.


[Without event filter]

Before (without cache)
15.8s, 333MB

Before (with cache)
 6.3s, 335MB

After
18.5s, 102MB

[With event filter]
Before (without cache)
13.2s, 274MB

Before(with cache)
3.84s, 245MB

After
13.9 s, 23MB